modify

+++ lightning_attn_cute/light_attn_without_decay.cu
@@ -108,7 +108,7 @@ __global__ void flash_forward(const half_t* q, const half_t* k, const half_t* v,
   const int bs_head_offset = bx * N * kHeadDim;
   const int num_block = (N + BLOCK - 1) / BLOCK;
 
-  __shared__ half_t smem_kv[kHeadDim * kHeadDim];
+  __shared__ float smem_kv[kHeadDim * kHeadDim];
 //  for (int i = tx; i < kHeadDim * kHeadDim; i += blockDim.x) {
 //    smem_kv[i] = __float2half(0.0f);
 //  }
